fix baggageclaim tests on windows
we needed to use the specific error from windows. Both
`ERROR_FILE_NOT_FOUND` and `ERROR_PATH_NOT_FOUND` fall under Go's
generic `os.ErrNotExist` error.

On macOS and Linux, the existing code works as expected.
